--- a/deep_gemm/include/deep_gemm/utils.cuh
+++ b/deep_gemm/include/deep_gemm/utils.cuh
@@ -46,3 +46,8 @@ template <typename T>
 __device__ __host__ constexpr T ceil_div(T a, T b) {
     return (a + b - 1) / b;
 }
+
+template <typename T>
+__device__ __host__ constexpr T gcd(T a, T b) {
+    return b == 0 ? a : gcd(b, a % b);
+}