By leveraging Large BlockTile optimization to alleviate **L2 cache pressure** and **maximize data reuse**, the H800-SXM achieves peak FP8 compute performance of 1470+ TFLOPS.
@LyricZhao 

### Normal GEMMs for dense models
|  M   |   N   |   K   | Base BMxBN  | Computation     | Opti BMxBN | Computation    | Speedup |
|:----:|:-----:|:-----:|:-----:      |:-----------:    |:-----:     |:-----------:   |:-------:|
| 4096 | 24576 | 1536  | 128x128     | 999  TF         | 128x160    | 1166 TF        |  16.72% |
| 4096 | 32768 |  512  | 128x128     | 591  TF         | 128x160    | 748  TF        |  **26.57**% |
| 4096 | 7168  | 16384 | 128x128     | 1404 TF         | 128x160    | **1470** TF        |  4.70%  |
| 4096 | 7168  | 2048  | 128x128     | 1031 TF         | 128x160    | 1204 TF        |  16.78% |

### Grouped GEMMs for MoE models (contiguous layout)
|  Groups  |  M   |   N   |   K   | Base BMxBN  | Computation     | Opti BMxBN | Computation    | Speedup |
|:----:    |:----:|:-----:|:-----:|:-----:      |:-----------:    |:-----:     |:-----------:   |:-------:|
| 4        | 8192 | 4096  | 7168  | 128x128     | 1317 TF         | 128x160    | 1381 TF        |  4.86%  |
| 4        | 8192 | 7168  | 2048  | 128x128     | 1114 TF         | 128x160    | 1262 TF        |  13.29% |
| 8        | 4096 | 4096  | 7168  | 128x128     | 1317 TF         | 128x160    | 1383 TF        |  5.01%  |
| 8        | 4096 | 7168  | 2048  | 128x128     | 1107 TF         | 128x160    | 1259 TF        |  13.73% |

Test on “H800”-SXM && CUDA 12.8.1