@LyricZhao 2025-04-07
Introduced a bug (wrong TMA multicast condition for grouped contiguous GEMM) in this PR, fixed in https://github.com/deepseek-ai/DeepGEMM/commit/b4ecf9c3ffbb571b691f24a3b7661edd475565b1.